requirements.txt
flask
gunicorn
numpy
pillow
requests
onnx
onnxruntime

